refactor: re itemized Base64
